requirements-contrib.txt
Flask-SQLAlchemy>=0.16
Flask-Redis>=0.3.0